pyproject.toml
[tool.poetry]
name = "twig-server"
version = "0.1.0"
license = "GNU-AGPLv3"
readme = "README.md"
description = "Server backing TWIG"
authors = ["Robert <xiurobert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.9"
Flask = "^2.2.2"
neo4j = "^4.4.5"
python-dotenv = "^0.20.0"
neomodel = "^4.0.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
